Move hc-sandbox wrapper to its own module
